+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0 (2016/12/02)
+
+- Add extraFiles property to config. This property copy extra files on documentation ouput directory.
+- Update process.cwd() path for grunt load in other modules
+
 ## 2.0.0 (2016/12/02)
 
 - Clean global workflow